ingestion/requirements.txt
Django==6.0.3
channels==4.2.2
daphne==4.2.1
celery[redis]==5.4.0
djangorestframework==3.15.2
drf-spectacular
psycopg[binary]
